refactor: refatorando o slide
