+++ src/pages/Home/containers/OurActions/Desktop/Carousel/Carousel.js
