+++ src/pages/Home/containers/index.js
@@ -1,9 +1,17 @@
 import AboutUs from './AboutUs/AboutUs';
 import Partners from './Partners/Partners';
 import WhereAreWe from './WhereAreWe/WhereAreWe';
-import Slide from './OurActions/Mobile/Slide';
-import SliderData from './OurActions/Desktop/SliderData';
+import SlideMobile from './OurActions/Mobile/SlideMobile';
+import SliderDesktop from './OurActions/Desktop/SliderDesktop';
 import Instagram from './Instagram/Instagram';
 import Banner from './Banner/Banner';
 
-export { AboutUs, Partners, WhereAreWe, SliderData, Slide, Instagram, Banner };
+export {
+  AboutUs,
+  Partners,
+  WhereAreWe,
+  SliderDesktop,
+  SlideMobile,
+  Instagram,
+  Banner,
+};
